refactor(relay): fixturesを本番のcore型単位でファイル分割
fixtureファイル名 = 生成する型のcoreモジュール名（data_object /
subscription）という対応にし、fixtureの置き場所を規則化する。
